feat: add route for get common busy time
