+++ app/presentation/views/calendar_list.slim
