+++ config/secrets.yml
@@ -1,0 +1,20 @@
+---
+development:
+  API_URL: http://localhost:3000/api/v1
+  APP_URL: http://localhost:9292
+  SESSION_SECRET: some_secret
+  MSG_KEY: NJlmA5m+wQZLHPFn/yejnsBYu+iRWFRDlMpvbKxiWA4=
+  REDIS_URL: rediss://:pc7bf696a91508ff1a2f1eba18563c08f1682af25b5a1e9eaf07e902ce02d4e22@ec2-54-221-218-75.compute-1.amazonaws.com:6910
+
+test:
+  API_URL: http://localhost:3000/api/v1
+  APP_URL: http://localhost:9292
+  MSG_KEY: 88EpejsFWNjm2d36hCUs5jo5wN57fgf1z1OdW/t2s/M=
+  SESSION_SECRET: some_secret
+
+production:
+  API_URL: https://calendar-coordinator-api.herokuapp.com/api/v1
+  APP_URL: https://calendar-coordinator-web.herokuapp.com
+  SESSION_SECRET: some_secret
+  MSG_KEY: Lw0LB8E4UJ6bBvzwR9VHpasc+y/TYjpjPuM43jkjtV4=
+  REDIS_URL: rediss://:pc7bf696a91508ff1a2f1eba18563c08f1682af25b5a1e9eaf07e902ce02d4e22@ec2-54-221-218-75.compute-1.amazonaws.com:6910
